requirements.txt
# to install packages run: pip install -r requirements.txt

python-docx==1.1.0